created helper-file; modularized requests
