+++ frontend/static/js/component-editor.js
@@ -1,3 +1,5 @@
+import {post_request} from './helper.js';
+
 //Base url to distinguish between localhost and production environment
 const base_url = window.location.origin;
 
@@ -103,7 +105,7 @@ function getFeatures() {
 
             let div = document.createElement('div');
             div.className = 'control-area';
-            div.innerHTML = '<a href="#" data-wait="Bitte warten..." id="save-button" class="create-button w-button" onclick="saveComponent()">Speichern</a>';
+            div.innerHTML = '<a href="#" data-wait="Bitte warten..." id="save-button" class="create-button w-button" onclick="createEditComponent()">Speichern</a>';
 
             // Append element to document
             document.getElementById('metrics-input').appendChild(div);
@@ -156,33 +158,6 @@ function processComponentData(json_data) {
 }
 
 /**
- * This function sends a post request to the backend
- *
- * @param {string} endpoint: The endpoint to be referred to
- * @param {string} data_json: The JSON Object to be passed to the backend
- * @param {function} callback: The function to be executed with the response
- */
-
-function post_request(endpoint, data_json, callback) {
-    const base_url = window.location.origin;
-    let xhttp = new XMLHttpRequest();
-    xhttp.open("POST", base_url + endpoint, true);
-    xhttp.setRequestHeader("Content-Type", "application/json");
-
-    // Handle response of HTTP-request
-    xhttp.onreadystatechange = function () {
-        if (this.readyState === XMLHttpRequest.DONE && (this.status >= 200 && this.status < 300)) {
-            // Process response and show sum in output field
-            let json = JSON.parse(this.responseText);
-            callback(json);
-        }
-    }
-
-    // Send HTTP-request
-    xhttp.send(data_json);
-}
-
-/**
  * This function enables and disables the component metrics for user input
  *
  * @param {string} selected_category: The component category selected in the dropdown
@@ -230,7 +205,7 @@ function toggleSection(element) {
  * This function saves the data entered to the database by transmitting the data to the backend
  */
 
-function saveComponent() {
+function createEditComponent() {
     let metric_elements = document.getElementsByClassName('metric-input');
     let metrics = {};
     let text_replaced_flag = false; // Helper variable that indicates, whether or not a non quantitative metric input has been found and discarded
